Moved upgrades code to its own class files.  organized a new manager.
